refactor: scroll로직 useScrollToggle로 리팩토링 및 적용(#53)
