+++ itda-front/src/components/Products/Products.tsx
@@ -1,6 +1,7 @@
 import Header from "components/common/Header/Header";
 import Navigator from "./Navigator";
 import ProductList from "./ProductList";
+import TopButton from "components/common/TopButton";
 import S from "./ProductsStyles";
 const Products = () => {
   return (
@@ -13,6 +14,7 @@ const Products = () => {
       </S.Products.HeaderLayout>
       <S.Products.MainLayout>
         <ProductList />
+        <TopButton />
       </S.Products.MainLayout>
     </>
   );
